Expected Pool Components

Returns the elements expected to be contained in the analyse object depending on what analysis method was specified.

Usage

get_pool_components(x)

Arguments

x

Character name of the analysis method, must one of either "rubin", "jackknife", "bootstrap" or "bmlmi".